english/lab06_deployment/requirements.txt
boto3
botocore
awscli
opensearch-py
requests-aws4auth
pyyaml
retrying
pandas
strands-agents
strands-agents-tools
streamlit
pillow
bedrock-agentcore-starter-toolkit
pydantic